added users queries in model
